+++ app/Http/Controllers/SiteController.php
@@ -19,6 +19,8 @@ class SiteController extends Controller
 {
     public function index(Server $server): View
     {
+        $this->authorize('manage', $server);
+
         return view('sites.index', [
             'server' => $server,
             'sites' => $server->sites()->orderByDesc('id')->get(),
@@ -27,6 +29,8 @@ class SiteController extends Controller
 
     public function store(Server $server, Request $request): HtmxResponse
     {
+        $this->authorize('manage', $server);
+
         $site = app(CreateSite::class)->create($server, $request->input());
 
         Toast::success('Site created');
@@ -36,6 +40,8 @@ class SiteController extends Controller
 
     public function create(Server $server): View
     {
+        $this->authorize('manage', $server);
+
         return view('sites.create', [
             'server' => $server,
             'type' => old('type', request()->query('type', SiteType::LARAVEL)),
@@ -45,6 +51,8 @@ class SiteController extends Controller
 
     public function show(Server $server, Site $site, Request $request): View|RedirectResponse|HtmxResponse
     {
+        $this->authorize('manage', $server);
+
         if (in_array($site->status, [SiteStatus::INSTALLING, SiteStatus::INSTALLATION_FAILED])) {
             if ($request->hasHeader('HX-Request')) {
                 return htmx()->redirect(route('servers.sites.installing', [$server, $site]));
@@ -61,6 +69,8 @@ class SiteController extends Controller
 
     public function installing(Server $server, Site $site, Request $request): View|RedirectResponse|HtmxResponse
     {
+        $this->authorize('manage', $server);
+
         if (! in_array($site->status, [SiteStatus::INSTALLING, SiteStatus::INSTALLATION_FAILED])) {
             if ($request->hasHeader('HX-Request')) {
                 return htmx()->redirect(route('servers.sites.show', [$server, $site]));
@@ -77,6 +87,8 @@ class SiteController extends Controller
 
     public function destroy(Server $server, Site $site): RedirectResponse
     {
+        $this->authorize('manage', $server);
+
         app(DeleteSite::class)->delete($site);
 
         Toast::success('Site is being deleted');
